requirements.txt
boto3==1.40.73
botocore==1.40.73
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
urllib3==1.26.20
